+++ graphbrain/tests/hypergraph.py
@@ -120,20 +120,13 @@ class Hypergraph:
         self.hg.add('(says/Pd mary/Cp)')
         self.hg.add('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))')
         self.hg.add('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C) extra/C)')
-        self.assertEqual(list(self.hg.search('(* graphbrain/Cp *)')),
-                         [hedge('(is/Pd graphbrain/Cp great/C)')])
-        self.assertEqual(list(self.hg.search('(is/Pd graphbrain/Cp *)')),
-                         [hedge('(is/Pd graphbrain/Cp great/C)')])
+        self.assertEqual(list(self.hg.search('(* graphbrain/Cp *)')), [hedge('(is/Pd graphbrain/Cp great/C)')])
+        self.assertEqual(list(self.hg.search('(is/Pd graphbrain/Cp *)')), [hedge('(is/Pd graphbrain/Cp great/C)')])
         self.assertEqual(list(self.hg.search('(x * *)')), [])
-        self.assertEqual(
-            list(
-                self.hg.search('(says/Pd * '
-                               '(is/Pd graphbrain/Cp great/C))')),
-            [hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))')])
-        self.assertEqual(
-            list(
-                self.hg.search('(says/Pd * (is/Pd * *))')),
-            [hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))')])
+        self.assertEqual(list(self.hg.search('(says/Pd * (is/Pd graphbrain/Cp great/C))')),
+                         [hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))')])
+        self.assertEqual(list(self.hg.search('(says/Pd * (is/Pd * *))')),
+                         [hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))')])
 
     def test_search_open_ended(self):
         self.hg.destroy()
@@ -141,19 +134,13 @@ class Hypergraph:
         self.hg.add('(says/Pd mary/Cp)')
         self.hg.add('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))')
         self.hg.add('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C) extra/C)')
-        self.assertEqual(list(self.hg.search('(* graphbrain/Cp * ...)')),
-                         [hedge('(is/Pd graphbrain/Cp great/C)')])
-        self.assertEqual(
-            list(self.hg.search('(is/Pd graphbrain/Cp * ...)')),
-            [hedge('(is/Pd graphbrain/Cp great/C)')])
+        self.assertEqual(list(self.hg.search('(* graphbrain/Cp * ...)')), [hedge('(is/Pd graphbrain/Cp great/C)')])
+        self.assertEqual(list(self.hg.search('(is/Pd graphbrain/Cp * ...)')), [hedge('(is/Pd graphbrain/Cp great/C)')])
         self.assertEqual(list(self.hg.search('(x * * ...)')), [])
 
-        self.assertEqual(
-            list(
-                self.hg.search(
-                    '(says/Pd * (is/Pd graphbrain/Cp great/C) ...)')),
-            [hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))'),
-             hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C) extra/C)')])
+        self.assertEqual(list(self.hg.search('(says/Pd * (is/Pd graphbrain/Cp great/C) ...)')),
+                         [hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))'),
+                          hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C) extra/C)')])
 
     def test_search_star(self):
         self.hg.destroy()
@@ -162,15 +149,10 @@ class Hypergraph:
         self.hg.add('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))')
         self.hg.add('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C) extra/C)')
         self.assertEqual(set(self.hg.search('*')),
-                         {hedge('(is/Pd graphbrain/Cp great/C)'),
-                          hedge('(says/Pd mary/Cp)'),
-                          hedge('(says/Pd mary/Cp '
-                                '(is/Pd graphbrain/Cp great/C))'),
-                          hedge('(says/Pd mary/Cp '
-                                '(is/Pd graphbrain/Cp great/C) extra/C)'),
-                          hedge('extra/C'), hedge('graphbrain/Cp'),
-                          hedge('great/C'), hedge('is/Pd'), hedge('mary/Cp'),
-                          hedge('says/Pd')})
+                         {hedge('(is/Pd graphbrain/Cp great/C)'), hedge('(says/Pd mary/Cp)'),
+                          hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))'),
+                          hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C) extra/C)'), hedge('extra/C'),
+                          hedge('graphbrain/Cp'), hedge('great/C'), hedge('is/Pd'), hedge('mary/Cp'), hedge('says/Pd')})
 
     def test_search_at(self):
         self.hg.destroy()
@@ -179,8 +161,7 @@ class Hypergraph:
         self.hg.add('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))')
         self.hg.add('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C) extra/C)')
         self.assertEqual(set(self.hg.search('.')),
-                         {hedge('extra/C'), hedge('graphbrain/Cp'),
-                          hedge('great/C'), hedge('is/Pd'), hedge('mary/Cp'),
+                         {hedge('extra/C'), hedge('graphbrain/Cp'), hedge('great/C'), hedge('is/Pd'), hedge('mary/Cp'),
                           hedge('says/Pd')})
 
     def test_search_amp(self):
@@ -190,62 +171,36 @@ class Hypergraph:
         self.hg.add('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))')
         self.hg.add('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C) extra/C)')
         self.assertEqual(set(self.hg.search('(*)')),
-                         {hedge('(is/Pd graphbrain/Cp great/C)'),
-                          hedge('(says/Pd mary/Cp)'),
-                          hedge('(says/Pd mary/Cp '
-                                '(is/Pd graphbrain/Cp great/C))'),
-                          hedge('(says/Pd mary/Cp '
-                                '(is/Pd graphbrain/Cp great/C) extra/C)')})
+                         {hedge('(is/Pd graphbrain/Cp great/C)'), hedge('(says/Pd mary/Cp)'),
+                          hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C))'),
+                          hedge('(says/Pd mary/Cp (is/Pd graphbrain/Cp great/C) extra/C)')})
 
     def test_search_non_atomic_pred(self):
         self.hg.destroy()
-        edge = hedge('((is/M playing/Pd ) mary/Cp.s '
-                     '(a/Md ((very/M old/Ma) violin/Cn.s)))')
+        edge = hedge('((is/M playing/Pd ) mary/Cp.s (a/Md ((very/M old/Ma) violin/Cn.s)))')
         self.hg.add(edge)
-        self.assertEqual(list(self.hg.search('((is/M playing/Pd) ...)')),
-                         [edge])
-        self.assertEqual(list(self.hg.search('((is/M playing/Pd) * *)')),
-                         [edge])
-        self.assertEqual(list(self.hg.search('((is/M playing/Pd) . (*))')),
-                         [edge])
-        self.assertEqual(list(self.hg.search('((is/M playing/Pd) . .)')),
-                         [])
-        self.assertEqual(list(self.hg.search('((is/M playing/Pd) (*) (*))')),
-                         [])
-        self.assertEqual(list(self.hg.search('(* mary/Cp.s *)')),
-                         [edge])
-        self.assertEqual(list(self.hg.search('(mary/Cp.s * *)')),
-                         [])
-        self.assertEqual(list(self.hg.search('(* * (a/Md ((very/M old/Ma) '
-                                             'violin/Cn.s)))')),
-                         [edge])
-        self.assertEqual(list(self.hg.search('((a/Md ((very/M old/Ma) '
-                                             'violin/Cn.s)) * *)')),
-                         [])
+        self.assertEqual(list(self.hg.search('((is/M playing/Pd) ...)')), [edge])
+        self.assertEqual(list(self.hg.search('((is/M playing/Pd) * *)')), [edge])
+        self.assertEqual(list(self.hg.search('((is/M playing/Pd) . (*))')), [edge])
+        self.assertEqual(list(self.hg.search('((is/M playing/Pd) . .)')), [])
+        self.assertEqual(list(self.hg.search('((is/M playing/Pd) (*) (*))')), [])
+        self.assertEqual(list(self.hg.search('(* mary/Cp.s *)')), [edge])
+        self.assertEqual(list(self.hg.search('(mary/Cp.s * *)')), [])
+        self.assertEqual(list(self.hg.search('(* * (a/Md ((very/M old/Ma) violin/Cn.s)))')), [edge])
+        self.assertEqual(list(self.hg.search('((a/Md ((very/M old/Ma) violin/Cn.s)) * *)')), [])
 
     def test_search_non_atomic_pred_vars(self):
         self.hg.destroy()
-        edge = hedge('((is/M playing/Pd ) mary/Cp.s '
-                     '(a/Md ((very/M old/Ma) violin/Cn.s)))')
+        edge = hedge('((is/M playing/Pd ) mary/Cp.s (a/Md ((very/M old/Ma) violin/Cn.s)))')
         self.hg.add(edge)
-        self.assertEqual(list(self.hg.search('((is/M playing/Pd) *X *Y)')),
-                         [edge])
-        self.assertEqual(list(self.hg.search('((is/M playing/Pd) .X (Y))')),
-                         [edge])
-        self.assertEqual(list(self.hg.search('((is/M playing/Pd) .X .Y)')),
-                         [])
-        self.assertEqual(list(self.hg.search('((is/M playing/Pd) (X) (Y))')),
-                         [])
-        self.assertEqual(list(self.hg.search('(* mary/Cp.s X)')),
-                         [edge])
-        self.assertEqual(list(self.hg.search('(mary/Cp.s X Y)')),
-                         [])
-        self.assertEqual(list(self.hg.search('(X Y (a/Md ((very/M old/Ma) '
-                                             'violin/Cn.s)))')),
-                         [edge])
-        self.assertEqual(list(self.hg.search('((a/Md ((very/M old/Ma) '
-                                             'violin/Cn.s)) X Y)')),
-                         [])
+        self.assertEqual(list(self.hg.search('((is/M playing/Pd) *X *Y)')), [edge])
+        self.assertEqual(list(self.hg.search('((is/M playing/Pd) .X (Y))')), [edge])
+        self.assertEqual(list(self.hg.search('((is/M playing/Pd) .X .Y)')), [])
+        self.assertEqual(list(self.hg.search('((is/M playing/Pd) (X) (Y))')), [])
+        self.assertEqual(list(self.hg.search('(* mary/Cp.s X)')), [edge])
+        self.assertEqual(list(self.hg.search('(mary/Cp.s X Y)')), [])
+        self.assertEqual(list(self.hg.search('(X Y (a/Md ((very/M old/Ma) violin/Cn.s)))')), [edge])
+        self.assertEqual(list(self.hg.search('((a/Md ((very/M old/Ma) violin/Cn.s)) X Y)')), [])
 
     def test_search_pred_with_roles(self):
         self.hg.destroy()
@@ -256,23 +211,25 @@ class Hypergraph:
 
     def test_search_pred_with_roles_and_ellipsis(self):
         self.hg.destroy()
-        edge = hedge(('(is/Mv.|f--3s-/en playing/Pd.so.|pg----/en)',
-                      'mary/Cp.s/en',
+        edge = hedge(('(is/Mv.|f--3s-/en playing/Pd.so.|pg----/en)', 'mary/Cp.s/en',
                       '(a/Md/en ((very/M/en old/Ma/en) violin/Cc.s/en))'))
         pattern = hedge('((is/Mv.|f--3s-/en playing/Pd.so.|pg----/en) ...)')
         self.hg.add(edge)
-        self.assertEqual(list(self.hg.search(pattern)),
-                         [edge])
+        self.assertEqual(list(self.hg.search(pattern)), [edge])
 
     def test_search_pred_with_deep_pattern(self):
         self.hg.destroy()
-        edge = hedge(('(is/Mv.|f--3s-/en playing/Pd.so.|pg----/en)',
-                      'mary/Cp.s/en',
+        edge = hedge(('(is/Mv.|f--3s-/en playing/Pd.so.|pg----/en)', 'mary/Cp.s/en',
                       '(a/Md/en ((very/M/en old/Ma/en) violin/Cc.s/en))'))
         pattern = hedge('(*/Pd.so */C (*/M ((*/M */M) violin/Cc.s/en)))')
         self.hg.add(edge)
-        self.assertEqual(list(self.hg.search(pattern)),
-                         [edge])
+        self.assertEqual(list(self.hg.search(pattern)), [edge])
+
+    def test_search_pred_with_multiple_incomplete_atoms(self):
+        self.hg.destroy()
+        self.hg.add('(says/Pd.so.|f--3s-/en mary/Cp hello/Cs)')
+        self.assertEqual(list(self.hg.search('(says/Pd.{so} mary/C hello/C)')),
+                         [hedge('(says/Pd.so.|f--3s-/en mary/Cp hello/Cs)')])
 
     def test_count(self):
         self.hg.destroy()
@@ -283,8 +240,7 @@ class Hypergraph:
         self.assertEqual(self.hg.count('(* graphbrain/Cp *)'), 1)
         self.assertEqual(self.hg.count('(is/Pd graphbrain/Cp *)'), 1)
         self.assertEqual(self.hg.count('(x * *)'), 0)
-        self.assertEqual(self.hg.count('(says/Pd * '
-                                       '(is/Pd graphbrain/Cp great/C))'), 1)
+        self.assertEqual(self.hg.count('(says/Pd * (is/Pd graphbrain/Cp great/C))'), 1)
         self.assertEqual(self.hg.count('(says/Pd * (is/Pd * *))'), 1)
 
     def test_count_open_ended(self):
@@ -401,7 +357,7 @@ class Hypergraph:
                            [{'X': hedge('(is/M playing/Pd)'),
                              'Y': hedge('mary/Cp.s')}])])
 
-    def test_match_argroles(self):
+    def test_match_argroles1(self):
         self.hg.destroy()
         edge = hedge('((is/M playing/Pd.so) mary/Cp.s '
                      '(a/Md ((very/M old/Ma) violin/Cn.s)))')
@@ -427,7 +383,7 @@ class Hypergraph:
               [{'X': hedge('mary/Cp.s'),
                 'Y': hedge('(a/Md ((very/M old/Ma) violin/Cn.s))')}])])
 
-    def test_match_argroles_curly_brackets(self):
+    def test_match_argroles2(self):
         self.hg.destroy()
         edge1 = hedge('(is/Pd.cs blue/Ca (the/M sky/C))')
         edge2 = hedge('(is/Pd.sc (the/M sky/C) blue/Ca)')
@@ -436,15 +392,13 @@ class Hypergraph:
         self.hg.add(edge2)
 
         self.assertEqual(
-            list(self.hg.match('(is/P.{sc} OBJ/C PROP)')),
-            [(edge1,
-              [{'OBJ': hedge('(the/M sky/C)'),
-                'PROP': hedge('blue/Ca')}]),
-             (edge2,
-              [{'OBJ': hedge('(the/M sky/C)'),
-                'PROP': hedge('blue/Ca')}])])
+            list(self.hg.match('(is/Pd.sc OBJ/C blue/Ca)')),
+            [(edge2, [{'OBJ': hedge('(the/M sky/C)')}])])
+        self.assertEqual(
+            list(self.hg.match('(is/P.sc OBJ/C blue/C)')),
+            [(edge2, [{'OBJ': hedge('(the/M sky/C)')}])])
 
-    def test_match_argroles_strict(self):
+    def test_match_argroles_curly_brackets1(self):
         self.hg.destroy()
         edge1 = hedge('(is/Pd.cs blue/Ca (the/M sky/C))')
         edge2 = hedge('(is/Pd.sc (the/M sky/C) blue/Ca)')
@@ -452,27 +406,19 @@ class Hypergraph:
         self.hg.add(edge1)
         self.hg.add(edge2)
 
-        self.assertEqual(
-            list(self.hg.match('(is/Pd.sc OBJ/C blue/Ca)', strict=True)),
-            [(edge2, [{'OBJ': hedge('(the/M sky/C)')}])])
-        self.assertEqual(
-            list(self.hg.match('(is/P.sc OBJ/C blue/C)', strict=True)),
-            [])
+        self.assertEqual(list(self.hg.match('(is/P.{sc} OBJ/C PROP)')),
+                         [(edge1, [{'OBJ': hedge('(the/M sky/C)'), 'PROP': hedge('blue/Ca')}]),
+                          (edge2, [{'OBJ': hedge('(the/M sky/C)'), 'PROP': hedge('blue/Ca')}])])
 
-    def test_match_argroles_not_strict(self):
+    def test_match_argroles_curly_brackets2(self):
         self.hg.destroy()
-        edge1 = hedge('(is/Pd.cs blue/Ca (the/M sky/C))')
-        edge2 = hedge('(is/Pd.sc (the/M sky/C) blue/Ca)')
+        edge1 = hedge('(is/Pd.cs blue/Ca (the/Md sky/Cc))')
+        edge2 = hedge('(is/Pd.sc (the/Md sky/Cc) blue/Ca)')
 
         self.hg.add(edge1)
         self.hg.add(edge2)
 
-        self.assertEqual(
-            list(self.hg.match('(is/Pd.sc OBJ/C blue/Ca)', strict=False)),
-            [(edge2, [{'OBJ': hedge('(the/M sky/C)')}])])
-        self.assertEqual(
-            list(self.hg.match('(is/P.sc OBJ/C blue/C)', strict=False)),
-            [(edge2, [{'OBJ': hedge('(the/M sky/C)')}])])
+        self.assertEqual(list(self.hg.match('(is/P.{sc} (the/M sky/C) blue/C)')), [(edge1, [{}]), (edge2, [{}])])
 
     def test_match_fun_patterns(self):
         self.hg.destroy()
@@ -484,25 +430,19 @@ class Hypergraph:
 
         self.assertEqual(
             list(self.hg.match('((atoms is/P.{sc}) (var */C OBJ) (var * PROP))')),
-            [(edge1,
-              [{'OBJ': hedge('(the/M sky/C)'),
-                'PROP': hedge('blue/Ca')}]),
-             (edge2,
-              [{'OBJ': hedge('(the/M sky/C)'),
-                'PROP': hedge('blue/Ca')}])])
+            [(edge1, [{'OBJ': hedge('(the/M sky/C)'), 'PROP': hedge('blue/Ca')}]),
+             (edge2, [{'OBJ': hedge('(the/M sky/C)'), 'PROP': hedge('blue/Ca')}])])
 
     def test_star(self):
         self.hg.destroy()
         edge1 = hedge('(is graphbrain/1 great/1)')
         self.hg.add(edge1)
-        self.assertEqual(list(self.hg.star(hedge('graphbrain/1'))),
-                         [edge1])
+        self.assertEqual(list(self.hg.star(hedge('graphbrain/1'))), [edge1])
         self.assertEqual(list(self.hg.star(hedge('graphbrain/2'))), [])
         self.hg.remove(edge1)
         edge2 = hedge('(says mary/1 (is graphbrain/1 great/1))')
         self.hg.add(edge2)
-        self.assertEqual(list(self.hg.star(edge1)),
-                         [edge2])
+        self.assertEqual(list(self.hg.star(edge1)), [edge2])
 
     def test_star_limit(self):
         self.hg.destroy()
@@ -518,11 +458,9 @@ class Hypergraph:
     def test_atoms_with_root(self):
         self.hg.destroy()
         self.hg.add('(is graphbrain/1 great/1)')
-        self.assertEqual(list(self.hg.atoms_with_root('graphbrain')),
-                         [hedge('graphbrain/1')])
+        self.assertEqual(list(self.hg.atoms_with_root('graphbrain')), [hedge('graphbrain/1')])
         self.hg.add('(is graphbrain/2 great/1)')
-        self.assertEqual(list(self.hg.atoms_with_root('graphbrain')),
-                         [hedge('graphbrain/1'), hedge('graphbrain/2')])
+        self.assertEqual(list(self.hg.atoms_with_root('graphbrain')), [hedge('graphbrain/1'), hedge('graphbrain/2')])
 
     def test_edges_with_edges(self):
         self.hg.destroy()
@@ -530,16 +468,9 @@ class Hypergraph:
         self.hg.add(edge1)
         edge2 = hedge('(is graphbrain/1 great/2)')
         self.hg.add(edge2)
-        self.assertEqual(
-            list(self.hg.edges_with_edges((hedge('graphbrain/1'),), 'great')),
-            [edge1, edge2])
-        self.assertEqual(list(self.hg.edges_with_edges((hedge('graphbrain/1'),
-                                                        hedge('is')),
-                                                       'great')),
-                         [edge1, edge2])
-        self.assertEqual(
-            list(self.hg.edges_with_edges((hedge('graphbrain/1'),), 'grea')),
-            [])
+        self.assertEqual(list(self.hg.edges_with_edges((hedge('graphbrain/1'),), 'great')), [edge1, edge2])
+        self.assertEqual(list(self.hg.edges_with_edges((hedge('graphbrain/1'), hedge('is')), 'great')), [edge1, edge2])
+        self.assertEqual(list(self.hg.edges_with_edges((hedge('graphbrain/1'),), 'grea')), [])
 
     def test_edges_with_edges2(self):
         self.hg.destroy()
@@ -549,17 +480,13 @@ class Hypergraph:
 
         res = self.hg.edges_with_edges([hedge('(of/B city/C lights/C)')])
         res = set([edge.to_str() for edge in res])
-        self.assertEqual(res, {'(syns/P (of/B city/C lights/C) paris/C)',
-                               '(going/P i/C (of/B city/C lights/C))'})
+        self.assertEqual(res, {'(syns/P (of/B city/C lights/C) paris/C)', '(going/P i/C (of/B city/C lights/C))'})
 
-        res = self.hg.edges_with_edges([hedge('(of/B city/C lights/C)'),
-                                        hedge('paris/C')])
+        res = self.hg.edges_with_edges([hedge('(of/B city/C lights/C)'), hedge('paris/C')])
         res = set([edge.to_str() for edge in res])
         self.assertEqual(res, {'(syns/P (of/B city/C lights/C) paris/C)'})
 
-        res = self.hg.edges_with_edges([hedge('(of/B city/C lights/C)'),
-                                        hedge('paris/C')],
-                                       'syns')
+        res = self.hg.edges_with_edges([hedge('(of/B city/C lights/C)'), hedge('paris/C')], 'syns')
         res = set([edge.to_str() for edge in res])
         self.assertEqual(res, {'(syns/P (of/B city/C lights/C) paris/C)'})